JBPM-6385: Split shortcuts for BPMN and DMN
